requirements.txt
boto3
pydantic
python-dotenv
